[FEAT] unified top menu generation
Previously every controller generated the top menu separately. The only
difference in the menus among controllers is which element is marked as
active. Therefore, a big amount of code was duplicated. Furthermore,
this made it hard to add new elements. This patch replaces all separate
autohandler templates with functionality in the already existing generic
autohandler template. Furthermore, new functionality in the existing
generic autohandler controller function populates the top menu in the
template and correctly sets the active element. Controllers can still
modify the top menu if needed.