[FIX] remove warning of odd hash elements
When showing a testplan or testrun by id the prepare_filter_path does
not work right. Thats ok, there are no filters in this case so it should
not even try to. When it does, the number of arguments is odd and
therefore does not make a hash. We therefore get a warning. I added a
check to get rid of this warning.