Unify navi generation
Long time ago every part in the Web GUI had its own top navi and side
navi template. With unification of top navis the side navi always used
the template of "Testruns". This template did not work for all parts of
the Web GUI because it expects data from the controller. Now all parts
generate the data for this template which in turn does no longer reside
in the testruns hierarchy. This means all parts of the Web GUI use the
same side navi template file and generate the required data.