catapult BindLex into outer space
Catalyst::Controller::BindLex allowed the
transparent mapping of a variable to the
Stash or Flash via it's name.

The syntax was nice using attributes but
the implementation behind it fragile which
is why it was deprecated years ago.

Now all usage of BindLex is gone in our code
which now unfortunately looks a bit
uglier because I wanted the replacement
to be as canonical as possible in this patchset.