[FEAT] show parse errors differently in reports overview
In the reports overview, reports with broken TAP are hard to recognise. They
get a success value of "FAIL" but a success ratio of 100% (all green). This
does not make sense. You can not trust the success ratio when there are parse
errors. Therefore, we now show these reports as grey.