Add machine name to metainfo overview
We found out by experience that sometimes it is easier to relate
metainfo in the "Test Execution Context" to the actual test when we also
provide the host name. This happens for example when you run the same
test on multiple guests of the same host. Then having the host name of
the guest machine can give better info.