[FIX] showing preconditions as valid YAML
In various places we show the preconditions associated with a
testrun. If a testrun has multiple preconditions they were concatenated
together. This usually worked because all preconditions entered into the
database come from YAML modules that stores them with a leading
'---'. Yet, this is not required and in fact unti tests sometimes use
YAML without leading '---' which is valid too.

To handle such a cases (and other weird cases where concatenation is not
enough) correctly we put all preconditions into one datastructure and
dump that out. This is always valid.