[FEAT] update preconditions in WebGUI
Changing a know wrong precondition with the CLI is not really easy. It
is far more comfortable to do this in the WebGUI. Therefore, this patch
offers a way to change the preconditions associated with testrun.

Note: This only makes sense in states 'schedule' or 'prepare' because
you can not change what a test looks like once its started (and
certainly not once its finished). This patch does not check this.