fix 'successgrade' in report list
The SELECT from the database returns
not 'successgrade' but 'r.successgrade'
which the HTML/Mason templates didn't
know yet.

Probably at a later point in history
this will get fixed or the SELECT
returns it without the prefix but
so far we just use 'r.successgrade'
when accessing the hashes in the templates.