replace MySQL-specific ALL()
ALL() is like a junction over sub-SELECTs.

Fortunately here it is used with the '>=' comparison,
so we can replace it with a simple MAX() inside the
sub-SELECT which has the same meaning.

It is probably less efficient than the optimized
ALL() that MySQL provides and also only works
with operators that can be expressed with
comparisons against aggregations, like here
the '>=' and MAX(), or '<=' and MIN().