testplan - Cancel
We now provide a link to Cancel a testplan.

This replaces the massively wrong testplan 'delete'
which just deleted here and there a bit everywhere,
just not something actually useful, so we ended up
with missing testruns, still running processes in
background, etc.

This patch relies on other fixes coming in the next minute.